State: Move isActivatingJetpackModule to global selectors directory
