+++ client/state/selectors/index.js
@@ -35,8 +35,9 @@ export getTimezonesByContinent from './get-timezones-by-continent';
 export getUpcomingBillingTransactions from './get-upcoming-billing-transactions';
 export hasBrokenSiteUserConnection from './has-broken-site-user-connection';
 export isActivatingJetpackJumpstart from './is-activating-jetpack-jumpstart';
-export isDeactivatingJetpackJumpstart from './is-deactivating-jetpack-jumpstart';
+export isActivatingJetpackModule from './is-activating-jetpack-module';
 export isAutomatedTransferActive from './is-automated-transfer-active';
+export isDeactivatingJetpackJumpstart from './is-deactivating-jetpack-jumpstart';
 export isDomainOnlySite from './is-domain-only-site';
 export isJetpackModuleActive from './is-jetpack-module-active';
 export isJetpackSiteInDevelopmentMode from './is-jetpack-site-in-development-mode';
